Treat 'prefix' like the correct variable type
The standard input files for grub2-mkconfig contain a variable 'prefix'
that gets set during compilation. The variable is then used by
grub2-mkconfig to populate several files in the configuration file.

The shell script references it as if it should still be a variable in
the finished product. This leads to grub mistakenly thinking that
'prefix' exists in the grub environment when it is actually a shell
variable set in 00_header.

Fixed the files containing references to 'prefix' to treat it like a
shell variable to be interpolated during the grub2-mkconfig run as
opposed to a grub environment variable.